backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.6

# Read .env file
pydantic-settings==2.4.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
alembic==1.13.3

# Vector database
chromadb==1.5.5

# HTTP client (for calling LLM API)
httpx==0.27.0

# File upload support
python-multipart==0.0.24

# RAG Pipeline — Text Extraction
PyMuPDF==1.27.2.2       # Extract text from PDF files (non-table content)
python-docx==1.2.0      # Extract text from DOCX files
pdfplumber==0.11.4      # Table extraction from PDF files

sentence-transformers==5.3.0